+++ src/components/Accounts/Actions/Actions.js
@@ -334,7 +334,7 @@ class Actions extends React.Component {
           )}
         </FormattedMessage>
         <CancellationModal
-          open={this.props.actions.cancellation}
+          open={actions.cancellation}
           onClose={this.onCloseCancellation}
           user={this.props.user}
           stripes={this.props.stripes}
@@ -342,16 +342,16 @@ class Actions extends React.Component {
           onSubmit={(values) => { this.onClickCancellation(values); }}
         />
         <PayModal
-          open={this.props.actions.pay || (this.props.actions.regular && !warning)}
+          open={actions.pay || (actions.regular && !warning)}
           commentRequired={settings.paid}
           onClose={this.onClosePay}
           account={[this.type]}
           stripes={this.props.stripes}
           balance={this.props.balance}
-          accounts={(this.props.actions.pay) ? this.props.accounts : accounts}
+          accounts={(actions.pay) ? this.props.accounts : accounts}
           payments={payments}
           onSubmit={(values) => {
-            if (this.props.actions.pay) {
+            if (actions.pay) {
               this.onClickPay(values);
             } else {
               this.onPayMany(values);
@@ -359,20 +359,20 @@ class Actions extends React.Component {
           }}
         />
         <WaiveModal
-          open={this.props.actions.waiveModal || (this.props.actions.waiveMany && !warning)}
+          open={actions.waiveModal || (actions.waiveMany && !warning)}
           commentRequired={settings.waived}
           onClose={this.onCloseWaive}
           stripes={this.props.stripes}
-          accounts={(this.props.actions.waiveModal) ? this.props.accounts : accounts}
+          accounts={(actions.waiveModal) ? this.props.accounts : accounts}
           balance={this.props.balance}
           waives={waives}
           onSubmit={(values) => {
-            if (this.props.actions.waiveModal) this.onClickWaive(values);
+            if (actions.waiveModal) this.onClickWaive(values);
             else this.onWaiveMany(values);
           }}
         />
         <CommentModal
-          open={this.props.actions.comment}
+          open={actions.comment}
           stripes={this.props.stripes}
           onClose={this.onCloseComment}
           onSubmit={(values) => { this.onClickComment(values); }}